Revise methods for cutting plane acc. comments (Part 1)
1.  row_coefficients should allow the row to be specified either as "a
string, an actual variable, or an integer interpreted as the index of a
variable", just like LPAbstractDictionary.leave() allows.
(See the examples in the docstring of enter and leave)

2. Since now I have the more general function row_coefficients, I can
implement "w" in terms of this function; in fact, this could take place
in the LPAbstractDictionary (remove leaving_coefficients from
LPDictionary and LPRevisedDictionary)

3. add_row actually has no output. Reword accordingly

4. In the tests of LPDictionary.add_row, replace the numbers with
decimal points by fractions, to avoid creating floating point numbers.

5. some of the matrix construction code can be simplified by forming a
matrix rather than a vector for the new row, and then using the "stack"
method of matrices

6. See
http://stackoverflow.com/questions/16730339/python-add-item-to-the-tuple
 on how to make tuples longer (make changes in add_row)

7. Instead of using QQ explicitly, it's better to use self.base_ring()
to find out what is the appropriate field