revise integer_variables, allow user to choose integer slack variables
One can infer integrality of the slack variables at setup time.

The sufficient conditions for a slack variable to be integer:
1. all problem variables are integer;
and 2. their coefficients for that row are integer;
and 3. the constant for that row is integer.

One should also allow the user to give the names of slack variables
that are known to be integer, even though the sufficient condition is
violated.